requirements.txt
flasgger==0.9.5
flask==2.2.2
flask-cors==3.0.10
importlib-metadata==4.12.0
itsdangerous==2.1.2
python-dotenv==0.21.0
Werkzeug==2.2.2
zipp==3.8.1
